chore(gpu): host-memory Phase-1 hygiene — H3 device memset, H4 u32 scatter indices, sll/sr fused tests
R4.3 per the round-3 spec, closing host-memory-workstream Phase 1:
- H3: the shard byte/range histograms are zeroed with a stream-ordered device
  memset instead of uploading a ~2 MB host zero-vec per shard.
- H4: host_lookup_scatter indices are u32 end-to-end (kernel, trait, RiscvAir
  impl, jagged caller) — max index < 2^18, halving the scatter upload.
  Behavior-neutral: histograms/scatter cells are bit-identical (existing
  fused-kernel tests pin them).
- sll/sr fused-entry direct tests (F.1b review note): device==CPU full-trace
  equality through generate_trace_device_with_lookups, fresh seeds,
  non-power-of-two counts so the device template fill is exercised on the
  fused path. Suite 99 -> 101.

Verified: sp1-gpu-tracegen 101/101, jagged 7/7, fmt/clippy clean, e2e
compressed prove verified.

+++ sp1-gpu/crates/jagged_tracegen/src/lib.rs
@@ -1138,14 +1138,14 @@ pub async fn ensure_trace_buffer_capacity(buffer: &mut PinnedBuffer<Felt>, requi
 /// the subsequent `hist_to_trace` (same stream) sees the complete counts. No-op if empty.
 fn scatter_into_histogram(
     hist: &mut DeviceBuffer<u32>,
-    idxs: &[u64],
+    idxs: &[u32],
     mults: &[u32],
     backend: &TaskScope,
 ) {
     if idxs.is_empty() {
         return;
     }
-    let mut idx_dev: Buffer<u64, TaskScope> = Buffer::with_capacity_in(idxs.len(), backend.clone());
+    let mut idx_dev: Buffer<u32, TaskScope> = Buffer::with_capacity_in(idxs.len(), backend.clone());
     idx_dev.extend_from_host_slice(idxs).unwrap();
     let mut mult_dev: Buffer<u32, TaskScope> =
         Buffer::with_capacity_in(mults.len(), backend.clone());
